intelligence/requirements.txt
androguard>=4.1.1
apkid>=2.1.5
pyyaml>=6.0
requests>=2.31.0
